requirements.txt
APScheduler==3.10.1
bson==0.5.10
certifi==2023.5.7
charset-normalizer==3.1.0
dateparser==1.1.8
dnspython==2.3.0
idna==3.4
oauthlib==3.2.2
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.14.0
tzlocal==5.0.1
urllib3==2.0.3
